refactor: Split subqueries.rs into smaller focused modules
Split the monolithic subqueries.rs (1001 lines) into logical submodules:
- cache.rs: Cache key computation for subquery results
- correlation.rs: Correlation detection and value extraction
- schema_utils.rs: Schema validation utilities
- scalar.rs: Scalar subquery evaluation
- exists.rs: EXISTS predicate evaluation
- quantified.rs: ALL/ANY/SOME comparisons
- in_subquery.rs: IN predicate evaluation with index optimization
- mod.rs: Module coordination and re-exports

Benefits:
- Improved code organization and maintainability
- Separation of concerns (caching, correlation, evaluation)
- Easier to navigate and understand specific subquery types
- Reduced cognitive load when working on specific features
- No changes to public API or functionality

Closes #2458
