+++ crates/vibesql-executor/src/evaluator/combined/subqueries/schema_utils.rs
@@ -1,0 +1,72 @@
+//! Schema validation utilities for subqueries
+//!
+//! This module provides utilities to compute and validate the schema of
+//! subquery results, particularly for handling wildcards and column counts.
+
+use crate::errors::ExecutorError;
+
+/// Compute the number of columns in a SELECT statement's result
+/// Handles wildcards by expanding them using table schemas from the database
+pub(super) fn compute_select_list_column_count(
+    stmt: &vibesql_ast::SelectStmt,
+    database: &vibesql_storage::Database,
+) -> Result<usize, ExecutorError> {
+    let mut count = 0;
+
+    for item in &stmt.select_list {
+        match item {
+            vibesql_ast::SelectItem::Wildcard { .. } => {
+                // Expand * to count all columns from all tables in FROM clause
+                if let Some(from) = &stmt.from {
+                    count += count_columns_in_from_clause(from, database)?;
+                } else {
+                    // SELECT * without FROM is an error (should be caught earlier)
+                    return Err(ExecutorError::UnsupportedFeature(
+                        "SELECT * requires FROM clause".to_string(),
+                    ));
+                }
+            }
+            vibesql_ast::SelectItem::QualifiedWildcard { qualifier, .. } => {
+                // Expand table.* to count columns from that specific table
+                let tbl = database
+                    .get_table(qualifier)
+                    .ok_or_else(|| ExecutorError::TableNotFound(qualifier.clone()))?;
+                count += tbl.schema.columns.len();
+            }
+            vibesql_ast::SelectItem::Expression { .. } => {
+                // Each expression contributes one column
+                count += 1;
+            }
+        }
+    }
+
+    Ok(count)
+}
+
+/// Count total columns in a FROM clause (handles joins and multiple tables)
+fn count_columns_in_from_clause(
+    from: &vibesql_ast::FromClause,
+    database: &vibesql_storage::Database,
+) -> Result<usize, ExecutorError> {
+    match from {
+        vibesql_ast::FromClause::Table { name, .. } => {
+            let table = database
+                .get_table(name)
+                .ok_or_else(|| ExecutorError::TableNotFound(name.clone()))?;
+            Ok(table.schema.columns.len())
+        }
+        vibesql_ast::FromClause::Join { left, right, .. } => {
+            let left_count = count_columns_in_from_clause(left, database)?;
+            let right_count = count_columns_in_from_clause(right, database)?;
+            Ok(left_count + right_count)
+        }
+        vibesql_ast::FromClause::Subquery { .. } => {
+            // For subqueries in FROM, we'd need to execute them to know column count
+            // This is complex, so for now we'll return an error
+            // In practice, this case is rare in IN subqueries
+            Err(ExecutorError::UnsupportedFeature(
+                "Subqueries in FROM clause within IN predicates are not yet supported for schema validation".to_string(),
+            ))
+        }
+    }
+}
